+++ src/Symfony/Component/Security/Core/Tests/Authorization/TraceableAccessDecisionManagerTest.php
@@ -16,31 +16,163 @@ use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use Symfony\Component\Security\Core\Authorization\AccessDecisionManager;
 use Symfony\Component\Security\Core\Authorization\DebugAccessDecisionManager;
 use Symfony\Component\Security\Core\Authorization\TraceableAccessDecisionManager;
+use Symfony\Component\Security\Core\Authorization\Voter\TraceableVoter;
+use Symfony\Component\Security\Core\Authorization\Voter\VoterInterface;
 
 class TraceableAccessDecisionManagerTest extends TestCase
 {
     /**
      * @dataProvider provideObjectsAndLogs
      */
-    public function testDecideLog($expectedLog, $object)
+    public function testDecideLog(array $expectedLog, array $attributes, $object, array $voterVotes, bool $result)
     {
-        $adm = new TraceableAccessDecisionManager(new AccessDecisionManager());
-        $adm->decide($this->getMockBuilder(TokenInterface::class)->getMock(), array('ATTRIBUTE_1'), $object);
+        $token = $this->createMock(TokenInterface::class);
 
-        $this->assertSame($expectedLog, $adm->getDecisionLog());
+        $admMock = $this
+            ->getMockBuilder(AccessDecisionManager::class)
+            ->setMethods(['decide'])
+            ->getMock();
+
+        $adm = new TraceableAccessDecisionManager($admMock);
+
+        $admMock
+            ->expects($this->once())
+            ->method('decide')
+            ->with($token, $attributes, $object)
+            ->willReturnCallback(function() use ($voterVotes, $adm, $result) {
+                foreach ($voterVotes as $voterVote) {
+                    list($voter, $vote) = $voterVote;
+                    $adm->addVoterVote($voter, $vote);
+                }
+
+                return $result;
+            })
+        ;
+
+        $adm->decide($token, $attributes, $object);
+
+        $this->assertEquals($expectedLog, $adm->getDecisionLog());
     }
 
-    public function provideObjectsAndLogs()
+    public function provideObjectsAndLogs(): \Generator
     {
-        $object = new \stdClass();
+        $voter1 = $this->getMockForAbstractClass(VoterInterface::class);
+        $voter2 = $this->getMockForAbstractClass(VoterInterface::class);
 
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => null, 'result' => false)), null);
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => true, 'result' => false)), true);
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => 'jolie string', 'result' => false)), 'jolie string');
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => 12345, 'result' => false)), 12345);
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => $x = fopen(__FILE__, 'r'), 'result' => false)), $x);
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => $x = array(), 'result' => false)), $x);
-        yield array(array(array('attributes' => array('ATTRIBUTE_1'), 'object' => $object, 'result' => false)), $object);
+        yield array(
+            array(array(
+                'attributes' => array('ATTRIBUTE_1'),
+                'object' => null,
+                'result' => true,
+                'voterDetails' => array(
+                    array('voter' => $voter1, 'vote' => VoterInterface::ACCESS_GRANTED),
+                    array('voter' => $voter2, 'vote' => VoterInterface::ACCESS_GRANTED),
+                ),
+            )),
+            array('ATTRIBUTE_1'),
+            null,
+            array(
+                array($voter1, VoterInterface::ACCESS_GRANTED),
+                array($voter2, VoterInterface::ACCESS_GRANTED)
+            ),
+            true
+        );
+        yield array(
+            array(array(
+                'attributes' => array('ATTRIBUTE_1'),
+                'object' => true,
+                'result' => false,
+                'voterDetails' => array(
+                    array('voter' => $voter1, 'vote' => VoterInterface::ACCESS_ABSTAIN),
+                    array('voter' => $voter2, 'vote' => VoterInterface::ACCESS_GRANTED),
+                ),
+            )),
+            array('ATTRIBUTE_1'),
+            true,
+            array(
+                array($voter1, VoterInterface::ACCESS_ABSTAIN),
+                array($voter2, VoterInterface::ACCESS_GRANTED)
+            ),
+            false
+        );
+        yield array(
+            array(array(
+                'attributes' => array(null),
+                'object' => 'jolie string',
+                'result' => false,
+                'voterDetails' => array(
+                    array('voter' => $voter1, 'vote' => VoterInterface::ACCESS_ABSTAIN),
+                    array('voter' => $voter2, 'vote' => VoterInterface::ACCESS_DENIED),
+                ),
+            )),
+            array(null),
+            'jolie string',
+            array(
+                array($voter1, VoterInterface::ACCESS_ABSTAIN),
+                array($voter2, VoterInterface::ACCESS_DENIED)
+            ),
+            false
+        );
+        yield array(
+            array(array(
+                'attributes' => array(12),
+                'object' => 12345,
+                'result' => true,
+                'voterDetails' => array(),
+            )),
+            'attributes' => array(12),
+            12345,
+            array(),
+            true
+        );
+        yield array(
+            array(array(
+                'attributes' => array(new \stdClass()),
+                'object' => $x = fopen(__FILE__, 'rb'),
+                'result' => true,
+                'voterDetails' => array(),
+            )),
+            array(new \stdClass()),
+            $x,
+            array(),
+            true
+        );
+        yield array(
+            array(array(
+                'attributes' => array('ATTRIBUTE_2'),
+                'object' => $x = array(),
+                'result' => false,
+                'voterDetails' => array(
+                    array('voter' => $voter1, 'vote' => VoterInterface::ACCESS_ABSTAIN),
+                    array('voter' => $voter2, 'vote' => VoterInterface::ACCESS_ABSTAIN),
+                ),
+            )),
+            array('ATTRIBUTE_2'),
+            $x,
+            array(
+                array($voter1, VoterInterface::ACCESS_ABSTAIN),
+                array($voter2, VoterInterface::ACCESS_ABSTAIN)
+            ),
+            false
+        );
+        yield array(
+            array(array(
+                'attributes' => array(12.13),
+                'object' => new \stdClass(),
+                'result' => false,
+                'voterDetails' => array(
+                    array('voter' => $voter1, 'vote' => VoterInterface::ACCESS_DENIED),
+                    array('voter' => $voter2, 'vote' => VoterInterface::ACCESS_DENIED),
+                ),
+            )),
+            array(12.13),
+            new \stdClass(),
+            array(
+                array($voter1, VoterInterface::ACCESS_DENIED),
+                array($voter2, VoterInterface::ACCESS_DENIED)
+            ),
+            false
+        );
     }
 
     public function testDebugAccessDecisionManagerAliasExistsForBC()